Edits for readability and style
Clarified that the hyphen in the 3rd position must be Y/N when CCPA applies. Providing the hyphen as an option was confusing, but clarified that it is only to be used when CCPA does not apply. Also removed periods in U.S. Just as CCPA has no periods, we don't need them in US either. And while "Digital Property" is a defined term in these documents, it's not a proper noun. I de-capitalized all instances. Bullets in the examples added to make the list of conditions easier to scan.